Improve code to convert a URL to a DCC accession.
The old function was only for the inclusion_list_url, but I extended
it to also work for the genome index. And since it was more general
compute_inclusion_id was renamed to
comute_dcc_file_accession_from_url.

The returned ID is in the normalized "/files/{accession}/" @id format,
and detects if it's passed the @id format the simple "{accession}"
format or an alias.
(Alias handling is not great and assumes that : won't be found in the
name.)