Try to detect if we managed to create the object before uploading
This checks if there the object has already been posted by md5sum and
if it has tries to grab the upload credentials from that before
uploading.

This might realistically replace the previous retry mechanism.

Open questions are what if the credentials expired? Should it try
to regenerate them.

Also sometimes it's annoying that snakemake deletes all your output
files.